[hotfix] Move test purpose example to e2e tests package in python.
