+++ python/flink_agents/e2e_tests/chat_model_example.py
