+++ python/flink_agents/e2e_tests/integrate_table_with_react_agent_example.py
@@ -32,7 +32,7 @@ from flink_agents.api.execution_environment import AgentsExecutionEnvironment
 from flink_agents.api.prompts.prompt import Prompt
 from flink_agents.api.resource import ResourceDescriptor
 from flink_agents.api.tools.tool import Tool
-from flink_agents.examples.common_tools import add, multiply
+from flink_agents.e2e_tests.common_tools import add, multiply
 from flink_agents.integrations.chat_models.ollama_chat_model import (
     OllamaChatModelConnection,
     OllamaChatModelSetup,
