+++ python/flink_agents/e2e_tests/resources/input_data.txt
